# football-tracker/docs/swagger.yaml
basePath: /
...
host: localhost
info:
  contact:
    email: support@football-tracker.com
    name: Football Tracker API Support
  description: REST API for tracking football championships, matches, and user predictions.
    Supports multi-championship tracking with user prediction system, rating, and
    leaderboard functionality.
  license:
    name: Proprietary
    url: '#'
  title: Football Tracker API
  version: "1.0"
paths:
...
  /api/v1/championships:
    get:
      description: Retrieves a list of all available football championships (FIFA
        World Cup, UEFA Euro, etc.)
      produces:
      - application/json
      responses:
        "200":
          description: List of championships
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all championships
      tags:
      - Championships
  /api/v1/championships/{id}:
    get:
      description: Retrieves detailed information about a specific championship
      parameters:
      - description: Championship ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Championship details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid championship ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get championship by ID
      tags:
      - Championships
...
schemes:
- https
- http
securityDefinitions:
  CookieAuth:
    description: Session token obtained from /api/v1/auth/login endpoint. After successful
      login, this cookie is automatically set.
    in: cookie
    name: session_token
    type: apiKey
swagger: "2.0"